Backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.42.0
anyio==3.7.1
asgiref==3.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
async-timeout==4.0.3
attrs==24.3.0
audioread==3.0.1
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.8
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
coloredlogs==15.0.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
cryptography==44.0.0
dataclasses-json==0.6.7
debugpy @ file:///D:/bld/debugpy_1734158945806/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1733236420667/work
deep-translator==1.11.4
defusedxml==0.7.1
Deprecated==1.2.15
distro==1.9.0
durationpy==0.9
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
fastapi==0.115.6
ffmpeg==1.4
filelock==3.16.1
filetype==1.2.0
firebase-admin==6.6.0
firecrawl-py==0.0.13
firestore==0.0.8
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
googletrans==4.0.0rc1
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.62.3
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.12.1
httpcore==0.9.1
httplib2==0.22.0
httptools==0.6.4
httpx==0.13.3
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/ipython_1701831845989/work
iso8601==2.1.0
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/jupyter_core_1727163532151/work
kubernetes==31.0.0
langchain==0.3.13
langchain-anthropic==0.1.15
langchain-chroma==0.1.4
langchain-community==0.3.13
langchain-core==0.3.28
langchain-google-firestore==0.3.0
langchain-google-genai==2.0.7
langchain-huggingface==0.1.2
langchain-openai==0.1.8
langchain-text-splitters==0.3.4
langchainhub==0.1.18
langdetect==1.0.9
langid==1.1.6
langsmith==0.1.147
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.58.1
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.12
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.0.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
pooch==1.8.2
posthog==3.7.4
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1733302527033/work
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.5
psutil @ file:///D:/bld/psutil_1729847171148/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pydub==0.25.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1733221634316/work
PyJWT==2.10.1
pyparsing==3.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==1.0.1
pywin32==307
PyYAML==6.0.2
pyzmq @ file:///D:/bld/pyzmq_1728642354337/work
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.9.4
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.13.1
sentence-transformers==3.0.0
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
SpeechRecognition==3.12.0
SQLAlchemy==2.0.36
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.41.3
sympy==1.13.1
tavily-python==0.3.3
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
tornado @ file:///D:/bld/tornado_1732615926917/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.47.1
typer==0.15.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.3
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.17.0
yarl==1.18.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
